requirements.txt
# NHL LiveStream Commentary Agent Dependencies
# Core ADK and AI
google-adk>=0.1.0
google-genai>=0.3.0
google-cloud-texttospeech>=2.16.0

# Web and API
requests>=2.31.0
websockets>=11.0

# Data Processing
python-dotenv>=1.0.0
pydantic>=2.0.0

# Audio Processing
pydub>=0.25.1
wave

# Async Support
asyncio
aiofiles>=23.0.0

# Development
pytest>=7.0.0